Refactored entire package to be import-friendly for third-party users
