+++ src/kineverse/__init__.py
@@ -1,29 +1,27 @@
-try:
-    import builtins
-    import inspect
+import kineverse.gradients.gradient_math as gm
+from kineverse.model.paths                 import Path, \
+                                                  PathDict, \
+                                                  PathSet, \
+                                                  PathException
+from kineverse.operations.operation     import Operation, \
+                                               OperationException
+from kineverse.operations.basic_operations import CreateValue, \
+                                                  ExecFunction
+from kineverse.operations.frame_operations import CreateRelativeFrame, \
+                                                  CreateRelativeTransform
+from kineverse.operations.urdf_operations  import load_urdf
+from kineverse.urdf_fix                    import load_urdf_file, \
+                                                  load_urdf_str
 
-    try:
-        f = profile
-    except NameError:
-        def profile(f):
-            return f
-        builtins.profile = profile
 
-    TYPE_CHECKING = True
-
-    def type_check(*types):
-        def dec_check(f):
-            if TYPE_CHECKING:
-                def checked_f(*args):
-                    args_start = 0 if next(iter(inspect.signature(f).parameters.keys())) != 'self' else 1
-                    if min([isinstance(v, t) for t, v in zip(types[:len(args) - args_start], args[args_start:])]) is False:
-                        raise Exception('Function {} called with wrong types:\n  Signature: {}\n      Given: {}'.format(str(f), ', '.join([str(t) for t in types[:len(args) - args_start]]), ', '.join([str(type(a)) for a in args[args_start:]])))
-                    return f(*args)
-                return checked_f
-            else:
-                return f
-        return dec_check
-
-    builtins.__dict__['type_check'] = type_check
-except:
-    print('Import of __builtin__ probably failed because this is run in Sphinx. Fingers crossed....')
+from kineverse.model.event_model    import EventModel
+from kineverse.model.geometry_model import GeometryModel, \
+                                           CollisionSubworld, \
+                                           ArticulatedObject, \
+                                           RigidBody, \
+                                           Box, \
+                                           Mesh, \
+                                           Cylinder, \
+                                           Sphere
+from kineverse.model.articulation_model import ArticulationModel, \
+                                               Constraint
