+++ src/kineverse/motion/__init__.py
@@ -1,0 +1,23 @@
+from .integrator import CommandIntegrator
+from .min_qp_builder import SoftConstraint, \
+                            ControlledValue, \
+                            MinimalQPBuilder as MQPB, \
+                            TypedQPBuilder   as TQPB, \
+                            PIDQPBuilder     as PQPB, \
+                            GeomQPBuilder    as GQPB, \
+                            PID_Constraint, \
+                            generate_controlled_values, \
+                            depth_weight_controlled_values, \
+                            find_constant_bounds
+
+from kineverse.model.geometry_model import create_distance_symbol, \
+                                           contact_geometry, \
+                                           closest_distance, \
+                                           closest_distance_constraint, \
+                                           contact_constraint, \
+                                           contact_geometry_world, \
+                                           closest_distance_world, \
+                                           closest_distance_constraint_world, \
+                                           contact_constraint_world, \
+                                           generate_contact_model
+
