+++ src/kineverse/ros/utils.py
@@ -1,0 +1,30 @@
+import os
+
+def res_pkg_path(rpath):
+    """Resolves a ROS package relative path to a global path.
+    :param rpath: Potential ROS URI to resolve.
+    :type rpath: str
+    :return: Local file system path
+    :rtype: str
+    """
+    if rpath[:10] == 'package://':
+        paths = os.environ['ROS_PACKAGE_PATH'].split(':')
+
+        rpath = rpath[10:]
+        pkg = rpath[:rpath.find('/')]
+
+        for rpp in paths:
+            if rpp[rpp.rfind('/') + 1:] == pkg:
+                return '{}/{}'.format(rpp[:rpp.rfind('/')], rpath)
+            if os.path.isdir('{}/{}'.format(rpp, pkg)):
+                return '{}/{}'.format(rpp, rpath)
+        raise Exception('Package "{}" can not be found in ROS_PACKAGE_PATH!'.format(pkg))
+    return rpath
+
+
+def make_pkg_path(path):
+    parts = path.split('/')[1:]
+    for x in reversed(range(len(parts) + 1)):
+        if os.path.isfile('/{}/package.xml'.format('/'.join(parts[:x]))):
+            return 'package://{}'.format('/'.join(parts[x-1:]))
+    return path
