+++ src/kineverse/utils.py
@@ -1,7 +1,6 @@
 import os
 
 import copy  as copy_module
-import importlib
 import math
 import numpy as np
 import kineverse.gradients.gradient_math as gm
@@ -13,51 +12,6 @@ if gm.cm.SYM_MATH_ENGINE == 'SYMENGINE':
     from kineverse.symengine_types         import symengine_types
 else:
     symengine_types = set()
-
-
-def res_pkg_path(rpath):
-    """Resolves a ROS package relative path to a global path.
-    :param rpath: Potential ROS URI to resolve.
-    :type rpath: str
-    :return: Local file system path
-    :rtype: str
-    """
-    if rpath[:10] == 'package://':
-        paths = os.environ['ROS_PACKAGE_PATH'].split(':')
-
-        rpath = rpath[10:]
-        pkg = rpath[:rpath.find('/')]
-
-        for rpp in paths:
-            if rpp[rpp.rfind('/') + 1:] == pkg:
-                return '{}/{}'.format(rpp[:rpp.rfind('/')], rpath)
-            if os.path.isdir('{}/{}'.format(rpp, pkg)):
-                return '{}/{}'.format(rpp, rpath)
-        raise Exception('Package "{}" can not be found in ROS_PACKAGE_PATH!'.format(pkg))
-    return rpath
-
-
-def make_pkg_path(path):
-    parts = path.split('/')[1:]
-    for x in reversed(range(len(parts) + 1)):
-        if os.path.isfile('/{}/package.xml'.format('/'.join(parts[:x]))):
-            return 'package://{}'.format('/'.join(parts[x-1:]))
-    return path
-
-
-def import_class(class_path):
-    """Imports a class using a type string.
-
-    :param class_path: Type string of the class.
-    :type  class_path: str
-    :rtype: type
-    """
-    components = class_path.split('.')
-    mod = importlib.import_module('.'.join(components[:-1]))
-    try:
-        return getattr(mod, components[-1])
-    except AttributeError:
-        raise AttributeError('Module "{}" loaded from "{}" has no attribute "{}"'.format(mod.__name__, '.'.join(components[:-1]), components[-1]))
 
 # Returns [x,y,z,w]
 RPY = namedtuple('RPY', ['r', 'p', 'y'])
